clarified the lack of need for the hash
The decryption should work fine if you remove the !!! and the hash following them. So long as you have the correct fernet key, you will still be able to decrypt. There will be warning messages that the hash doesn't match but you should still be able to decrypt. I clarified the print statements of the encryption function to more clearly represent this.